requirements.txt
fairseq @ git+https://github.com/pytorch/fairseq.git@fcca32258c8e8bcc9f9890bf4714fa2f96b6b3e1
hydra-colorlog==1.1.0
hydra-core==1.0.7
julius==0.2.6
omegaconf==2.0.6
opencv-python==4.5.4.60
pesq==0.0.3
pystoi==0.3.3
six==1.16.0
speechbrain==0.5.9
SoundFile==0.10.3.post1
torch
torchaudio
tensorboardX
pillow
tqdm==4.62.1
wandb